[KAN-75] refactor: /gatherings/[id] 페이지 컴포넌트 분리
